+++ internal/impl/pure/cache_ttlru_test.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/benthosdev/benthos/v4/public/service"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -15,22 +16,26 @@ func TestTTLRUCacheDefault(t *testing.T) {
 	defConf, err := ttlruCacheConfig().ParseYAML(``, nil)
 	require.NoError(t, err)
 
-	c, err := ttlruMemCacheFromConfig(defConf)
+	logger := service.MockResources().Logger()
+
+	c, err := ttlruMemCacheFromConfig(defConf, logger)
 	require.NoError(t, err)
 
 	testServiceCache(t, c)
 }
 
-func TestTTLRUCacheOptimisticAndWithoutReset(t *testing.T) {
+func TestTTLRUCacheOptimistic(t *testing.T) {
 	t.Parallel()
 
 	defConf, err := ttlruCacheConfig().ParseYAML(`
 optimistic: true
-without_reset: true
+without_reset: true # must be ignored
 `, nil)
 	require.NoError(t, err)
 
-	c, err := ttlruMemCacheFromConfig(defConf)
+	logger := service.MockResources().Logger()
+
+	c, err := ttlruMemCacheFromConfig(defConf, logger)
 	require.NoError(t, err)
 
 	testServiceCache(t, c)
@@ -40,7 +45,7 @@ func TestTTLRUCacheInitValues(t *testing.T) {
 	t.Parallel()
 
 	defConf, err := ttlruCacheConfig().ParseYAML(`
-ttl: '5m'
+default_ttl: '5m'
 cap: 1024
 init_values:
   foo: bar
@@ -48,7 +53,9 @@ init_values:
 `, nil)
 	require.NoError(t, err)
 
-	c, err := ttlruMemCacheFromConfig(defConf)
+	logger := service.MockResources().Logger()
+
+	c, err := ttlruMemCacheFromConfig(defConf, logger)
 	require.NoError(t, err)
 
 	ctx := context.Background()
@@ -72,7 +79,9 @@ func BenchmarkTTLRU(b *testing.B) {
 	defConf, err := ttlruCacheConfig().ParseYAML(``, nil)
 	require.NoError(b, err)
 
-	c, err := ttlruMemCacheFromConfig(defConf)
+	logger := service.MockResources().Logger()
+
+	c, err := ttlruMemCacheFromConfig(defConf, logger)
 	require.NoError(b, err)
 
 	ctx := context.Background()
@@ -95,7 +104,9 @@ func BenchmarkTTLRUParallel(b *testing.B) {
 	defConf, err := ttlruCacheConfig().ParseYAML(``, nil)
 	require.NoError(b, err)
 
-	c, err := ttlruMemCacheFromConfig(defConf)
+	logger := service.MockResources().Logger()
+
+	c, err := ttlruMemCacheFromConfig(defConf, logger)
 	require.NoError(b, err)
 
 	ctx := context.Background()