+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - Allow comments in single root and directly imported bloblang mappings
+- The `azure_blob_storage` input no longer adds `blob_storage_content_type` and `blob_storage_content_encoding` metadata values as string pointer types, and instead adds these values as string types only when they are present.
+- The `http_server` input now returns a more appropriate 503 service unavailable status code during shutdown instead of the previous 404 status.
 
 ## 4.20.0 - 2023-08-22